last login bug

+++ room8d/flmate/views.py
@@ -108,29 +108,32 @@ def edit(request):
             for you in profileList:
                 if not you.get('active'):
                     inactiveUser = User.objects.get(id=you.get('user_id'))
+
                     lastlog = inactiveUser.last_login
                     lastpos = timezone.now()- td(days=21)
-                    if lastlog<lastpos:
-                        for chat in Chatroom.objects.all():
-                            if inactiveUser in chat.members.all():
-                                chat.delete()
-                        inactiveUser.email_user("FLATMATE: До встречи!", 
-                                'Спасибо, ' + hUser.username +'!\nТы замечательный человек и мы уверены - ты нашел своего соседа! Уже 3 недели как ты не заходил на сайт.\nЕсли мы снова понадобимся тебе или твоим друзьям --> https://flatm8.ru/ \nДо новых встреч, дорогой друг!', 
-                                    'flatmate@flatm8.ru')
-                        inactiveUser.profile.delete()
-                        inactiveUser.delete()
+                    if lastlog:
+                        if lastlog<lastpos:
+                            for chat in Chatroom.objects.all():
+                                if inactiveUser in chat.members.all():
+                                    chat.delete()
+                            inactiveUser.email_user("FLATMATE: До встречи!", 
+                                    'Спасибо, ' + hUser.username +'!\nТы замечательный человек и мы уверены - ты нашел своего соседа! Уже 3 недели как ты не заходил на сайт.\nЕсли мы снова понадобимся тебе или твоим друзьям --> https://flatm8.ru/ \nДо новых встреч, дорогой друг!', 
+                                        'flatmate@flatm8.ru')
+                            inactiveUser.profile.delete()
+                            inactiveUser.delete()
                     continue
                 else:
                     UserToSwitch = User.objects.get(id=you.get('user_id'))
                     lastlog = UserToSwitch.last_login
-                    lastpos = timezone.now()- td(days=14)
-                    if lastlog<=lastpos:
-                        UserToSwitch.profile.active = False
-                        UserToSwitch.profile.save()
-                        UserToSwitch.email_user("FLATMATE - твой аккаунт деактивирован!", 
-                                'Привет, ' + hUser.username +'!\nСайт растет и число пользователей ежедневно увеличивается! Ты не заходил на сайт более двух недель и мы решили, что ты больше не ищешь соседа, поэтому деактивировали твой профиль. Если мы ошиблись - заходи на сайт и твой профиль снова активируется, иначе твой аккаунт будет безвозвратно удален через неделю! --> https://flatm8.ru/', 
-                                    'flatmate@flatm8.ru')
-                        continue    #IF EVERYTHING F*CK UP DELETE THIS LINE
+                    if lastlog:
+                        lastpos = timezone.now()- td(days=14)
+                        if lastlog<=lastpos:
+                            UserToSwitch.profile.active = False
+                            UserToSwitch.profile.save()
+                            UserToSwitch.email_user("FLATMATE - твой аккаунт деактивирован!", 
+                                    'Привет, ' + hUser.username +'!\nСайт растет и число пользователей ежедневно увеличивается! Ты не заходил на сайт более двух недель и мы решили, что ты больше не ищешь соседа, поэтому деактивировали твой профиль. Если мы ошиблись - заходи на сайт и твой профиль снова активируется, иначе твой аккаунт будет безвозвратно удален через неделю! --> https://flatm8.ru/', 
+                                        'flatmate@flatm8.ru')
+                            continue    #IF EVERYTHING F*CK UP DELETE THIS LINE
                          
                 if me == you or [me.get('user_id'),you.get('user_id')] in checked: continue
                 points = 0
